구간 관리 페이지 UI 구현 및 핸들러 함수 모듈화
