+++ index.html
@@ -6,7 +6,7 @@
   </head>
   <body>
     <div id="app">
-      <h1>🚇 지하철 노선도 관리 </h1>
+      <h1>🚇 지하철 노선도 관리</h1>
     </div>
     <script type="module" src="src/index.js"></script>
   </body>
