--- src/components/views/sectionManagerPage.js
@@ -1,5 +1,0 @@
-function sectionManagerPage() {
-  return `sectionManagerPage`;
-}
-
-export default sectionManagerPage;
